% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{print_the_data}
\alias{print_the_data}
\title{Print the simulated data}
\usage{
print_the_data()
}
\arguments{
\item{""}{}
}
\value{
a dataframe
}
\description{
Print the simulated data
}
\examples{
print_the_data()
}
